Added Product Backlog
Decided I wanted to start treating our project more like an actual development project and though we'll still need to meet to discuss our development method, I wanted to start creating a backlog of features/tasks that we need to work on. Feel free to add to it as much or as little as you wish.  For now it only servers the purpose of reminding us what we need to get done. My only request is finished tasks not be deleted, but instead moved to either a new file or a new sub section of the backlog.txt. This is for the autopsy after we finisht the projecct.